+++ src/App.jsx
@@ -6,6 +6,7 @@ import Mobile_landingPage from './pages/Mobile_LandingPage/Mobile_landingPage'
 import SignInPage_Mobile from './pages/SignInPage_Mobile/SignINPage_Mobile'
 import SignUpPageMobile from './pages/SignUpPageMobile/SignUpPageMobile'
 
+
 const ResponsiveComponent = ({ DesktopComponent, MobileComponent }) => {
   const [isMobile, setIsMobile] = useState(window.innerWidth < 768)
